Simplify thread delete handling
